pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "hatch-semver"
description = "Hatch plugin for semver versioning scheme"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "Sven Siegmund", email = "sven.siegmund@gmail.com" },
]

keywords = [
    "hatch",
    "plugin",
    "version",
    "development",
    "versioning",
]

classifiers = [
    #"Development Status :: 3 - Alpha",
    #"Development Status :: 4 - Beta",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
    "Framework :: Hatch",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.9",
    "Operating System :: OS Independent",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
]

dependencies = [
    "hatchling",
    "semver",
]

dynamic = [
    "version",
]

[project.urls]
Homepage = "https://fleetingbytes.github.io/hatch-semver/"
Repository = "https://github.com/fleetingbytes/hatch-semver"
Issues = "https://github.com/fleetingbytes/hatch-semver/issues"
PyPI = "https://pypi.org/project/hatch-semver/"

[project.entry-points.hatch]
semver = "hatch_semver.plugin.hooks"

[project.scripts]

[tool.hatch.version]
path = "src/hatch_semver/__about__.py"
validate-bump = true
scheme = "semver"

[tool.hatch.envs.default]
dependencies = [
    "pytest",
    "pytest-cov",
    "wheel",
    "towncrier",
    "twine",
]

[tool.hatch.envs.test]
dependencies = [
    "pytest",
    "pytest-cov",
    "hatch",
]

[tool.hatch.envs.default.scripts]
cov = "pytest -vx --cov-report=term-missing --cov-config=pyproject.toml --cov=src/hatch_semver --cov=tests"
no-cov = "cov --no-cov"

[tool.hatch.envs.style]
dependencies = [
    "black",
    "isort",
]

[tool.hatch.envs.style.scripts]
fmt = [
    "isort .",
    "black .",
]

[tool.hatch.envs.tc]
dependencies = [
    "towncrier",
]

[tool.hatch.envs.tc.scripts]
draft = [
    "towncrier build --draft",
    "git add --update",
    "git commit --message \"updated changelog\""
]
build = [
    "towncrier build --yes",
    # "Copy-Item -path ./CHANGELOG.md -destination ./docs -force",
    "copy /Y ./CHANGELOG.md ./docs",
    "git add --update",
    "git commit --message \"updated changelog\""
]

[tool.hatch.envs.docs]
dependencies = [
    "mkdocs",
    "mkdocs-material",
    "pdoc3",
]

[tool.hatch.envs.docs.scripts]
build = "pdoc --html --force --output-dir api-site hatch_semver"
serve = "pdoc --http : hatch_semver"
userve = [
    "mkdocs serve",
]

[[tool.hatch.envs.test.matrix]]
python = ["311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  #"src/hatch_semver/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.black]
    line-length = 102

[tool.towncrier]
name = "hatch-semver"
package = "hatch_semver"
package_dir = "src"
directory = "changelog.d"
filename = "CHANGELOG.md"
start_string = "<!-- towncrier release notes start -->\n"
underlines = ["", "", ""]
template = "changelog.d/changelog_template.jinja"
title_format = "## [{version}](https://github.com/fleetingbytes/hatch-semver/tree/{version}) - {project_date}"
issue_format = "[#{issue}](https://github.com/fleetingbytes/hatch-semver/issues/{issue})"
orphan_prefix = "+"

[tool.towncrier.fragment.doc]
name = "Documentation"
showcontent = true

[tool.towncrier.fragment.feature]
name = "New Features"
showcontent = true

[tool.towncrier.fragment.improved]
name = "Improvements"
showcontent = true

[tool.towncrier.fragment.fixed]
name = "Bugfixes"
showcontent = true

[tool.towncrier.fragment.unimportant]
name = "Development Details"
showcontent = true
